backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
python-dotenv==1.0.0
google-generativeai==0.3.0
aiohttp==3.11.10
pydantic==2.6.4
gitpython==3.1.40
aiofiles==23.2.1
gunicorn==21.2.0
